requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autograd==1.3
autograd-gamma==0.5.0
backcall==0.2.0
black==21.12b0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
fonttools==4.29.0
formulaic==0.2.4
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
google-auth==2.5.0
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
importlib-metadata==4.10.1
interface-meta==1.2.4
ipykernel==6.7.0
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
kiwisolver==1.3.2
lifelines==0.26.4
lightning-bolts==0.5.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==6.0.2
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.1
oauthlib==3.1.1
openpyxl==3.0.9
openslide-python==1.1.2
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.26
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-lightning==1.5.9
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sklearn==0.0
stack-data==0.1.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tomli==1.2.3
torch==1.10.2
torchmetrics==0.7.0
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
